fix(product-builds): deduplicate getCommitUrl utility
Extract getCommitUrl into shared formatting.js utility to eliminate
duplication across three view files.
